+++ src/concurrent_dll.h
@@ -20,7 +20,7 @@ int listInsert(struct DoubleLinkedList *dll, struct DLLNode *node);
 int listDelete(struct DoubleLinkedList *dll, int productID);
 int listSearch(struct DoubleLinkedList *dll, int productID);
 int DLLSize(struct DoubleLinkedList *dll);
-int DLLProductIdSum(struct DoubleLinkedList *dll);
+long long int DLLProductIdSum(struct DoubleLinkedList *dll);
 void printDLL(struct DoubleLinkedList *dll);
 
 #endif /*CONCURRENT_DLL_H*/
